+++ tests/projects/tracing_on_tests/+util/potato.m
@@ -1,0 +1,3 @@
+function x = potato(y)
+    x = y + 1;
+end