+++ tests/projects/tracing_on_tests/test/TestPotato.m
@@ -1,0 +1,10 @@
+classdef TestPotato < matlab.unittest.TestCase
+    
+    methods (Test, TestTags = {'Something'})
+        function test_01 (self)
+            self.verifyEqual(5, util.potato(4));
+        end
+    end
+    
+end
+